refactor: move external packages to internal directory
