+++ internal/gemini/client.go
